refator with base class
